Web console: fix tab switching stomping over each other (#1275)
* sample query progress

* update DQT

* fix query stomping

* cache tab promise

* fix start time

* slight cleanup

* better rate

* working with arrays

* update snapshots

* fix column ordering

* update tests

* fix up data loader

+++ web-console/src/views/workbench-view/helper-query/helper-query.tsx
@@ -25,8 +25,13 @@ import React, { useEffect, useRef } from 'react';
 import { Loader, QueryErrorPane } from '../../../components';
 import { usePermanentCallback, useQueryManager } from '../../../hooks';
 import { Api } from '../../../singletons';
+import { ExecutionStateCache } from '../../../singletons/execution-state-cache';
 import { WorkbenchHistory } from '../../../singletons/workbench-history';
-import { ColumnMetadata, DruidError, QueryAction, RowColumn } from '../../../utils';
+import {
+  WorkbenchRunningPromise,
+  WorkbenchRunningPromises,
+} from '../../../singletons/workbench-running-promises';
+import { ColumnMetadata, DruidError, QueryAction, QueryManager, RowColumn } from '../../../utils';
 import { QueryContext } from '../../../utils/query-context';
 import {
   DruidEngine,
@@ -40,7 +45,6 @@ import { ExecutionDetailsTab } from '../execution-details-pane/execution-details
 import { ExecutionErrorPane } from '../execution-error-pane/execution-error-pane';
 import { ExecutionProgressPane } from '../execution-progress-pane/execution-progress-pane';
 import { ExecutionStagesPane } from '../execution-stages-pane/execution-stages-pane';
-import { ExecutionStateCache } from '../execution-state-cache';
 import { ExecutionSummaryPanel } from '../execution-summary-panel/execution-summary-panel';
 import { ExecutionTimerPanel } from '../execution-timer-panel/execution-timer-panel';
 import {
@@ -95,12 +99,14 @@ export const HelperQuery = React.memo(function HelperQuery(props: HelperQueryPro
 
   const id = query.getId();
   const [executionState, queryManager] = useQueryManager<
-    WorkbenchQuery | LastExecution,
+    WorkbenchQuery | WorkbenchRunningPromise | LastExecution,
     Execution,
     Execution,
     DruidError
   >({
-    initQuery: ExecutionStateCache.getState(id) ? undefined : query.getLastExecution(),
+    initQuery: ExecutionStateCache.getState(id)
+      ? undefined
+      : WorkbenchRunningPromises.getPromise(id) || query.getLastExecution(),
     initState: ExecutionStateCache.getState(id),
     processQuery: async (q, cancelToken) => {
       if (q instanceof WorkbenchQuery) {
@@ -124,7 +130,8 @@ export const HelperQuery = React.memo(function HelperQuery(props: HelperQueryPro
           case 'sql': {
             if (cancelQueryId) {
               void cancelToken.promise
-                .then(() => {
+                .then(cancel => {
+                  if (cancel.message === QueryManager.TERMINATION_MESSAGE) return;
                   return Api.instance.delete(
                     `/druid/v2${engine === 'sql' ? '/sql' : ''}/${Api.encodePath(cancelQueryId)}`,
                   );
@@ -132,22 +139,35 @@ export const HelperQuery = React.memo(function HelperQuery(props: HelperQueryPro
                 .catch(() => {});
             }
 
+            onQueryChange(props.query.changeLastExecution(undefined));
+
             let result: QueryResult;
             try {
-              result = await queryRunner.runQuery({
+              const resultPromise = queryRunner.runQuery({
                 query,
                 extraQueryContext: mandatoryQueryContext,
-                cancelToken,
               });
+              WorkbenchRunningPromises.storePromise(id, { promise: resultPromise, sqlPrefixLines });
+
+              result = await resultPromise;
             } catch (e) {
-              onQueryChange(props.query.changeLastExecution(undefined));
               throw new DruidError(e, sqlPrefixLines);
             }
 
-            onQueryChange(props.query.changeLastExecution(undefined));
             return Execution.fromResult(engine, result);
           }
         }
+      } else if (WorkbenchRunningPromises.isWorkbenchRunningPromise(q)) {
+        let result: QueryResult;
+        try {
+          result = await q.promise;
+        } catch (e) {
+          WorkbenchRunningPromises.deletePromise(id);
+          throw new DruidError(e, q.sqlPrefixLines);
+        }
+
+        WorkbenchRunningPromises.deletePromise(id);
+        return Execution.fromResult('sql', result);
       } else {
         switch (q.engine) {
           case 'sql-task':
@@ -259,6 +279,7 @@ export const HelperQuery = React.memo(function HelperQuery(props: HelperQueryPro
             minimal
             onClick={() => {
               ExecutionStateCache.deleteState(id);
+              WorkbenchRunningPromises.deletePromise(id);
               onDelete();
             }}
           />
